# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Laravel Generator APIs",
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/role/addRole": {
            "post": {
                "tags": [
                    "Role & Permission"
                ],
                "summary": "Store a newly created Role in storage",
                "description": "Store Role",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "nasffme",
                        "in": "body",
                        "description": "Role that should be stored",
                        "required": false,
                        "schema": {
                            "$ref": "config('roles.models.role')"
                        }
                    },
                    {
                        "name": "bddddddddody",
                        "in": "body",
                        "description": "Role that should be stored",
                        "required": false,
                        "schema": {
                            "$ref": "config('roles.models.role')"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "config('roles.models.role')"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/role/addPermission": {
            "post": {
                "tags": [
                    "Role & Permission"
                ],
                "summary": "Store a newly created Permission in storage",
                "description": "Store Permission",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "in": "body",
                        "description": "Permission that should be stored",
                        "required": false,
                        "schema": {
                            "$ref": "config('Permissions.models.Permission')"
                        }
                    },
                    {
                        "name": "slug",
                        "in": "body",
                        "description": "Permission that should be stored",
                        "required": false,
                        "schema": {
                            "$ref": "config('Permissions.models.Permission')"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "config('Permissions.models.Permission')"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/role/addpermissiontorole": {
            "post": {
                "tags": [
                    "Role & Permission"
                ],
                "summary": "Add Permission to Role",
                "description": "Add Permission to Role",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "roleId",
                        "in": "body",
                        "description": "Role Id",
                        "required": false,
                        "schema": {
                            "$ref": "config('roles.models.role')"
                        }
                    },
                    {
                        "name": "permissionId",
                        "in": "body",
                        "description": "Permission Id",
                        "required": false,
                        "schema": {
                            "$ref": "config('permissions.models.permission')"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "config('roles.models.role')"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/role/deletepermission": {
            "post": {
                "tags": [
                    "Role & Permission"
                ],
                "summary": "delete Permission from Role",
                "description": "Store Role",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "roleId",
                        "in": "body",
                        "description": "Role Id",
                        "required": false,
                        "schema": {
                            "$ref": "config('roles.models.role')"
                        }
                    },
                    {
                        "name": "permissionId",
                        "in": "body",
                        "description": "Permission Id",
                        "required": false,
                        "schema": {
                            "$ref": "config('permission.models.permission')"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "all Permission should be deleted",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "config('roles.models.role')"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/role/deleteallpermission": {
            "post": {
                "tags": [
                    "Role & Permission"
                ],
                "summary": "delete all Permission",
                "description": "delete all Permission",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "body",
                        "description": "id role",
                        "required": false,
                        "schema": {
                            "$ref": "config('roles.models.role')"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "all Permission should be deleted",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "config('roles.models.role')"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
...
}